do not flush db when in production; do this manually in redis-cli

+++ src/services/jobQueue/index.ts
@@ -4,16 +4,18 @@ import { gdOrgReviewScraperJobQueueManager } from '../../GdOrgReviewRenewal/scra
 import { createClient } from 'redis';
 import { redisManager } from '../redis';
 import { s3OrgsJobQueueManager } from '../../GdOrgReviewRenewal/s3OrgsJob/queue';
+import { RuntimeEnvironment } from '../../utilities/runtime';
 
 export const SUPERVISOR_JOB_CONCURRENCY = 4;
 
 export const startJobQueues = () => {
     // TODO: add a if block once we add feature of resuming failed cronjob
-    // process.env.NODE_ENV === RuntimeEnvironment.DEVELOPMENT
-    const redisAdminClient = createClient(redisManager.config);
-    redisAdminClient.flushdb();
-    console.debug(`flushed redis db ${redisManager.config.db}`);
-    redisAdminClient.quit();
+    if (process.env.NODE_ENV !== RuntimeEnvironment.PRODUCTION) {
+        const redisAdminClient = createClient(redisManager.config);
+        redisAdminClient.flushdb();
+        console.debug(`flushed redis db ${redisManager.config.db}`);
+        redisAdminClient.quit();
+    }
 
     // register job queues
     // supervisorJobQueue.empty();
